feat: Add order delete auth check

+++ shop_springboot/src/main/java/com/example/shop/controller/OrderController.java
@@ -234,9 +234,14 @@ public class OrderController {
     return "result";
   }
 
-  @DeleteMapping(value = "/delete/id/{orderId}")
-  public void deleteOrder(@PathVariable String orderId) {
-     orderService.deleteOrder(orderId);
+  @DeleteMapping(value = "/delete")
+  public void deleteOrder(
+      @RequestParam String orderId,
+      @RequestParam String hashKey,
+      @RequestParam String id) {
+    if(adminService.checkAdmin(hashKey,id)) {
+      orderService.deleteOrder(orderId);
+    }
   }
 
 }
